backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
google-generativeai==0.4.1
email-validator==2.1.1
pydantic==2.6.4
pydantic-settings==2.2.1
sendgrid==6.11.0
